+++ src/components/App.css
@@ -1,0 +1,2 @@
+@import url(/src/shared/reset.css);
+@import url(/src/shared/common.css);
