+++ src/components/Cards.js
@@ -1,5 +1,7 @@
 import "./Cards.css";
 import noImg from "../assets/images/noImg.png";
+import favoritesImg from "../assets/images/star.svg";
+import meatballMenuImg from "../assets/images/meatball.svg";
 import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 dayjs.extend(relativeTime);
@@ -14,11 +16,13 @@ function Cards({ cardInfo }) {
     <li className="card-list">
       <a href={url} target="_blank" rel="noreferrer noopener">
         <div className="card-box">
-          <img className="card" src={imageSource ?? noImg} alt="카드이미지" />
+          <img className="card" src={imageSource ?? noImg} alt="카드 이미지" />
+          <img className="favorites" src={favoritesImg} alt="즐겨찾기 이미지" />
         </div>
         <div className="card-info">
           <div className="update-time">
             <p>{updateTimeAgo}</p>
+            <img src={meatballMenuImg} alt="추가메뉴 버튼" />
           </div>
           <div className="description">
             <p>{description}</p>
