+++ src/components/FolderInfo.css
