+++ src/components/GlobalStyle.js
@@ -1,0 +1,5 @@
+import { createGlobalStyle } from 'styled-components';
+
+const GlobalStyle = createGlobalStyle`
+
+`
