+++ src/pages/Folder.js
@@ -1,0 +1,16 @@
+import "./Folder.css";
+import Nav from "../components/Nav";
+import AddLink from "../components/AddLink";
+
+function Folder() {
+  return (
+    <>
+      <section className="folder-section">
+        <Nav />
+        <AddLink />
+      </section>
+    </>
+  );
+}
+
+export default Folder;
